+++ NEWS.md
@@ -1,3 +1,5 @@
-# ComprehensivePipelineUtils (development version)
+# ComprehensivePipelineUtils 0.0.0.9001
 
 * Initial CRAN submission.
+* Added a `NEWS.md` file to track changes to the package.
+- Add a new function `read_tcga_xena`.